Extract localRoute and router into separate files.
